+++ Source/Project64-core/N64System/MemoryHandler/RomMemoryHandler.h
@@ -38,7 +38,7 @@ private:
         _this->LoadedGameState();
     }
 
-    uint32_t & m_PC;
+    uint64_t & m_PC;
     CRegisters & m_Reg;
     CN64Rom & m_Rom;
     bool m_RomWrittenTo;
