+++ Source/Project64/UserInterface/Debugger/CPULog.cpp
@@ -53,7 +53,7 @@ void CCPULog::PushState()
 
     CPUState * state = &m_Array[m_Index++];
 
-    state->pc = g_Reg->m_PROGRAM_COUNTER;
+    state->pc = (uint32_t)g_Reg->m_PROGRAM_COUNTER;
     state->opcode = g_System->Opcode();
 
     memcpy(state->gpr, g_Reg->m_GPR, sizeof(g_Reg->m_GPR));
